[ 1429375 ] Uniformity for Annotation action values
Added String constants for ENABLE and RENDER to Action class and USER and ADMIN constants to Roles class.  We can add more common roles in the future, but the point is to only provide basic functionality here that helps people not screw up their string constants.  Typesafety is not possible with annotations.  A very secure project should not use them at all.  But it should be at least possible to make your annotations a little safer this way.  

Code now looks like this in the example:

@AuthorizeAction(action = Action.RENDER, roles = { Roles.ADMIN, Roles.USER })

It's not ideal, but then annotations are not ideal.


git-svn-id: https://svn.apache.org/repos/asf/incubator/wicket/trunk@459297 13f79535-47bb-0310-9956-ffa450edef68