Fix RatingsPanel demo
In short: this is the *easy* fix.

Problem: 'reset' links do not reset the ratings.
Details: the demo uses _static_ models which are never serialized. The reset links' models wraps the static ones and are serialized. After deserialization the wrapped are no more linked to the static ones and don't work anymore. Making the page non-versioned solves the problem without the need to re-work the demo.



git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@983303 13f79535-47bb-0310-9956-ffa450edef68