Add IHeaderResponse#renderCSSReference() and #renderJavaScriptReference() that accept PageParameters additionaly.
This way it is possible to create resource urls with query parameters.
This is used by the new resource decoration example.

The problem was that until now the application code created the URL with the parameters with urlFor(ResourceReference, PageParameters) and then pass it to IHeaderResponse#renderCSSReference(String) which again relativized the url and thus broke it (the same for Javascript).

With this commit the new example works as it was in the original application for 1.4.x

TODO: make HomePage extending WicketExamplePage so that it has the top banner and navigation.

git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@1051071 13f79535-47bb-0310-9956-ffa450edef68