WICKET-3670 WicketTester does not support custom pages for component testing using the new API WICKET-3699 BaseWicketTester startPanel(panel) missing
Rename startComponent(Class [, IMarkupFragment]) to startComponentInPage() so it is more clear that this component is added to a page to be tested in contrast to #startComponent(Component) which just renders the component.
Add factory methods for the page and its markup so the client code can customize them as needed.
Add better generics support in all these methods.



git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@1102010 13f79535-47bb-0310-9956-ffa450edef68