WICKET-4170 Dump the threads' stack traces if a thread cannot acquire the lock on a page instance
Manually dump the threads' stack traces because there is no API in Java 5/6 to do this automatically.
The format is close to the real thread dump format but not exact because the API doesn't provide all needed information (tid, nid, waiting conditions, ...).
The most important information is there: the name of the thread that cannot acquire the lock and the name of the thread which holds the lock.



git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@1189104 13f79535-47bb-0310-9956-ffa450edef68