WICKET-4601 IResourceStreamWriter to take OutputStream instead of Response
IResourceStreamWriter#write() now accepts java.io.OutputStream (as in Wicket 1.4) instead of o.a.w.Response

Move IResourceStreamWriter and AbstractResourceStreamWriter to wicket-util since they don't have dependencies to classes in -core anymore. This way their package name is the same as in 1.5.x (before the OSGi-friendly change made in 6.x).