WICKET-4915 org.apache.wicket.util.resource.Patht#find fails on Windows
Fix a failing test on Windows.
Path#find() expects a relative url against the folder it was initialized with.
The test is passing on Unix because new File("/", "/tmp/something") leads to "//tmp/something" and this is OK.
In Windows: new File("C:\", "C:\tmp\something") leads to "C:\C:\tmp\something".

patch-by: cleiter