WICKET-5531 Create new placeholder tag to indicate where header contributions should appear
Remove the usage of the tag namespace when checking where the tag should be transformed from openClose to open+close.
The namespace is dynamic (it depends) on the xmlns: for the page. We cannot hardcode namespaces in the list of tags to expand

(cherry picked from commit 66fd9d41b24ac4010542d75f4566bf9978f7728b)